requirements.txt
# Core
numpy>=1.23.0
scipy>=1.10.0

# Audio I/O
soundfile>=0.12.0
pyaudio>=0.2.11

# Web / streaming
websockets>=10.0
python-dotenv>=1.0.0

# RVC dependencies
fairseq>=0.12.2
faiss-cpu>=1.7.4

# NOTE: PyTorch is required but is NOT pinned here (install the correct build for your system):
#   https://pytorch.org/get-started/locally/
#
# Optional (file formats / extra f0 methods):
#   av (PyAV)  -> enables reading more formats via ffmpeg in some RVC utilities
#   pyworld    -> enables f0_method=harvest
#   torchcrepe -> enables f0_method=crepe
